+++ tests/CashierTest.php
@@ -350,7 +350,8 @@ class CashierTest extends TestCase
         ]);
 
         // Create Invoice
-        $user->createAsStripeCustomer($this->getTestToken());
+        $user->createAsStripeCustomer();
+        $user->updateCard($this->getTestToken());
         $user->invoiceFor('Laravel Cashier', 1000);
 
         // Invoice Tests
@@ -367,7 +368,8 @@ class CashierTest extends TestCase
         ]);
 
         // Create Invoice
-        $user->createAsStripeCustomer($this->getTestToken());
+        $user->createAsStripeCustomer();
+        $user->updateCard($this->getTestToken());
         $invoice = $user->invoiceFor('Laravel Cashier', 1000);
 
         // Create the refund